requirements.txt
streamlit
streamlit-authenticator
streamlit-lottie
requests
json
time
numpy
pandas
emojis
urlextract
matplotlib
wordcloud